+++ Plugins/Funapi/Source/Funapi/Private/google/protobuf/unknown_field_set.cc
@@ -54,7 +54,7 @@ UnknownFieldSet::~UnknownFieldSet() {
 
 void UnknownFieldSet::ClearFallback() {
   GOOGLE_DCHECK(fields_ != NULL);
-  for (int i = 0; i < fields_->size(); i++) {
+  for (int i = 0; i < (int)fields_->size(); i++) {
     (*fields_)[i].Delete();
   }
   fields_->clear();
@@ -78,7 +78,7 @@ int UnknownFieldSet::SpaceUsedExcludingSelf() const {
   if (fields_ == NULL) return 0;
 
   int total_size = sizeof(*fields_) + sizeof(UnknownField) * fields_->size();
-  for (int i = 0; i < fields_->size(); i++) {
+  for (int i = 0; i < (int)fields_->size(); i++) {
     const UnknownField& field = (*fields_)[i];
     switch (field.type()) {
       case UnknownField::TYPE_LENGTH_DELIMITED:
@@ -161,7 +161,7 @@ void UnknownFieldSet::DeleteSubrange(int start, int num) {
     (*fields_)[i + start].Delete();
   }
   // Slide down the remaining fields.
-  for (int i = start + num; i < fields_->size(); ++i) {
+  for (int i = start + num; i < (int)fields_->size(); ++i) {
     (*fields_)[i - num] = (*fields_)[i];
   }
   // Pop off the # of deleted fields.
@@ -173,7 +173,7 @@ void UnknownFieldSet::DeleteSubrange(int start, int num) {
 void UnknownFieldSet::DeleteByNumber(int number) {
   if (fields_ == NULL) return;
   int left = 0;  // The number of fields left after deletion.
-  for (int i = 0; i < fields_->size(); ++i) {
+  for (int i = 0; i < (int)fields_->size(); ++i) {
     UnknownField* field = &(*fields_)[i];
     if (field->number() == number) {
       field->Delete();
